+++ src/components/typewriter.tsx
